added auhentication check for deleting post

+++ controllers/api/v1/posts_api.js
@@ -21,12 +21,18 @@ module.exports.index = async function (req, res) {
 module.exports.destroy = async function (req, res) {
   try {
     const post = await Post.findById(req.params.id);
-    await Post.deleteOne({ _id: req.params.id });
-    await Comment.deleteMany({ post: req.params.id });
+    if (post.user == req.user.id) {
+      await Post.deleteOne({ _id: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    return res.json(200, {
-      message: "Post and associated comments deleted successfullt",
-    });
+      return res.json(200, {
+        message: "Post and associated comments deleted successfullt",
+      });
+    } else {
+      return res.json(401, {
+        message: "You cannot delete this post",
+      });
+    }
   } catch (err) {
     console.log("********", err);
     return res.json(500, {
